Statistic Assignment1 Q.3
Q.3) A car believes that the percentage of citizens in city ABC that owns a vehicle is 60% or less. A sales manager disagrees with this. He conducted a hypothesis testing surveying 250 residents & found that 170 residents responded yes to owning a vehicle.
(a) State the null & alternate hypothesis.
(b) At a 10% significance level, is there enough evidence to support the idea that vehicle owner in ABC city is 60% or less.